Improve update routine
Sometimes the application does fail to communicate with the GitLab
"Update" Server.
To prevent this I setup a proxy on my server (update.l0nax.org) so all
requests to a file will be forwarded only ONCE.
This prevents "crashes" with errors like "Timeout […]".

To respect the privacy of all 'changelog-go' users, the
'update.l0nax.org' proxy will not forward the request IP nor will the
requests be logged.

The privacy is also an aspect why the Update Proxy has been deployed.